requirements.txt
tensorflow
numpy
streamlit
pillow
opencv-python-headless
